Updated Python Version

+++ Codechef/Detscore/Detscore.py
@@ -3,5 +3,5 @@ while T:
     T -= 1
     X, N = map(int, input().split())
     X /= 10
-    score = N * X
+    score = int(N * X)
     print(score)
